+++ src/lib/pages/proposal-details/components/vote-details/deposit-period/depositors-table/index.tsx
@@ -1,5 +1,6 @@
 import { TableContainer } from "@chakra-ui/react";
 
+import { Loading } from "lib/components/Loading";
 import { EmptyState } from "lib/components/state";
 import type { ProposalDeposit } from "lib/types";
 
@@ -8,11 +9,24 @@ import { DepositorsTableRow } from "./DepositorsTableRow";
 
 interface DepositorsTableProps {
   depositors: ProposalDeposit[];
+  isDepositsLoading: boolean;
+  isPruned: boolean;
 }
 
 const templateColumns = `1.5fr 1.5fr 2fr 1fr`;
 
-export const DepositorsTable = ({ depositors }: DepositorsTableProps) => {
+export const DepositorsTable = ({
+  depositors,
+  isDepositsLoading,
+  isPruned,
+}: DepositorsTableProps) => {
+  if (isDepositsLoading) return <Loading />;
+
+  if (isPruned)
+    return (
+      <EmptyState message="The proposal has ended, and the depositor list is already pruned." />
+    );
+
   if (depositors.length === 0)
     return <EmptyState message="The proposal has no depositors yet." />;