+++ src/lib/services/types/proposal.ts
@@ -16,6 +16,7 @@ import type {
   Coin,
   Proposal,
   ProposalData,
+  ProposalDeposit,
   ProposalParams,
   ProposalStatus,
   ProposalValidatorVote,
@@ -46,7 +47,7 @@ export const zProposalParamsResponse = z
   .transform<ProposalParams<Coin>>(snakeToCamel);
 
 export const zProposalParamsResponseLcd = z.object({
-  param: zProposalParamsResponse,
+  params: zProposalParamsResponse,
 });
 
 export const zProposal = z.object({
@@ -97,9 +98,14 @@ export const zProposalDataResponse = z.object({
         .array(),
       resolved_timestamp: zUtcDate.nullable(),
       submit_time: zUtcDate,
-      total_deposit: zCoin.array().nullable(),
+      total_deposit: zCoin.array(),
       version: z.string(),
       voting_time: zUtcDate.nullable(),
+      yes: zBig,
+      abstain: zBig,
+      no: zBig,
+      no_with_veto: zBig,
+      resolved_total_voting_power: zBig.nullable(),
     })
     .transform<ProposalData<Coin>>(
       ({ created_tx_hash, proposal_deposits, messages, ...val }) => ({
@@ -112,6 +118,13 @@ export const zProposalDataResponse = z.object({
           txHash: parseTxHash(deposit.tx_hash),
         })),
         messages,
+        finalTallyResult: {
+          yes: val.yes,
+          abstain: val.abstain,
+          no: val.no,
+          noWithVeto: val.no_with_veto,
+          totalVotingPower: val.resolved_total_voting_power,
+        },
       })
     )
     .nullable(),
@@ -124,9 +137,35 @@ export const zProposalVotesInfoResponse = z
     abstain: zBig,
     no: zBig,
     no_with_veto: zBig,
-    total_voting_power: zBig.nullable(),
+    current_total_voting_power: zBig.nullable(),
   })
-  .transform<ProposalVotesInfo>(snakeToCamel);
+  .transform<ProposalVotesInfo>((val) => ({
+    ...snakeToCamel(val),
+    totalVotingPower: val.current_total_voting_power,
+  }));
+
+export const zProposalVotesInfoResponseLcd = z
+  .tuple([
+    z.object({
+      yes_count: zBig,
+      abstain_count: zBig,
+      no_count: zBig,
+      no_with_veto_count: zBig,
+    }),
+    z.object({
+      pool: z.object({
+        not_bonded_tokens: zBig,
+        bonded_tokens: zBig,
+      }),
+    }),
+  ])
+  .transform<ProposalVotesInfo>(([tally, pool]) => ({
+    yes: tally.yes_count,
+    abstain: tally.abstain_count,
+    no: tally.no_count,
+    noWithVeto: tally.no_with_veto_count,
+    totalVotingPower: pool.pool.bonded_tokens,
+  }));
 
 const zProposalVotesResponseItem = z
   .object({
@@ -185,10 +224,10 @@ export const zProposalDataResponseLcd = z
     messages: z.array(zMessageResponse).nullable(),
     status: z.string(),
     final_tally_result: z.object({
-      yes_count: z.string(),
-      no_count: z.string(),
-      abstain_count: z.string(),
-      no_with_veto_count: z.string(),
+      yes_count: zBig,
+      no_count: zBig,
+      abstain_count: zBig,
+      no_with_veto_count: zBig,
     }),
     submit_time: zUtcDate,
     deposit_end_time: zUtcDate,
@@ -203,6 +242,13 @@ export const zProposalDataResponseLcd = z
   })
   .transform<ProposalData<Coin>>((val) => ({
     ...snakeToCamel(val),
+    finalTallyResult: {
+      yes: val.final_tally_result.yes_count,
+      abstain: val.final_tally_result.abstain_count,
+      no: val.final_tally_result.no_count,
+      noWithVeto: val.final_tally_result.no_with_veto_count,
+      totalVotingPower: null,
+    },
     status: val.status
       .replace("PROPOSAL_STATUS_", "")
       .split("_")
@@ -218,8 +264,7 @@ export const zProposalDataResponseLcd = z
     description: val.summary,
     proposalDeposits: [],
     resolvedTimestamp: null,
-    version: "v1",
-    votingTime: null,
+    votingTime: val.voting_start_time,
   }));
 export type ProposalDataResponseLcd = z.infer<typeof zProposalDataResponseLcd>;
 
@@ -228,3 +273,15 @@ export const zProposalsResponseLcd = z.object({
   pagination: zPagination,
 });
 export type ProposalsResponseLcd = z.infer<typeof zProposalsResponseLcd>;
+
+export const zProposalDepositsResponseLcd = z.object({
+  deposits: z.array(
+    z
+      .object({
+        depositor: zBechAddr,
+        amount: zCoin.array(),
+      })
+      .transform<ProposalDeposit<Coin>>((deposit) => deposit)
+  ),
+  pagination: zPagination,
+});