+++ src/lib/utils/zod.ts
@@ -8,7 +8,11 @@ export const parseWithError = <T, S extends ZodTypeDef, U>(
     return zod.parse(data);
   } catch (e) {
     // eslint-disable-next-line no-console
-    console.error("Zod parsing error: ", (e as ZodError).errors);
+    console.error(
+      "Zod parsing error: ",
+      (e as ZodError).name,
+      (e as ZodError).errors
+    );
     throw e;
   }
 };